Rename timers folder to timer
