+++ src/components/game/mechanics/helpers/eventListener/addPauseDetection.test.js
@@ -1,10 +1,10 @@
 import EventListener from "./eventListener";
 import AudioManager from "../audio/audioManager";
-import Timer from "../timers/timer";
+import Timer from "../timer/timer";
 import Animator from "../graphics/animator/animator";
 
 jest.mock("../audio/audioManager");
-jest.mock("../timers/timer");
+jest.mock("../timer/timer");
 jest.mock("../graphics/animator/animator");
 
 let variables;
