+++ src/components/game/mechanics/helpers/timer/pauseTimers.test.js
