+++ src/components/game/mechanics/helpers/timer/timer.js
